subcommands: allow passing in tags for build ids
Where appropriate, allow passing a build tag (using "@") where a build
id is expected.  For example:

gbp status lighthouse 123

is allowed but also

gbp status lighthouse @prod

In the latter case, "@prod" is resolved using the GraphQL API. If the
tag does not exist for the given machine then an error is raised.