notes: don't open editor when reading from stdin
This bug has existed for a while. I guess until now I've never used `gbp
notes` from stdin but when I did I found that it would still open up my
editor.  The problem is that it's checking if stdout is a tty when it
should be checking for stdin.