machines: don't use * to denote published build
I didn't like using "*" to denote a published build as it overloaded the
symbol. In the `list` command it's used do denote a build that has
packages. But `P` looks like a bad a a prefix. This change instead
utilizes color (green) and weight (bold) to denote a published build.