diff: fix attribute error when diffing old builds
In the olden days of gbp Builds did not have a 'built' property.
The built property defaults to None and stays that way for old builds.
As such we need to treat them special.

diff shows dates for builds and uses the built timestamp in the output.
If built is None then we get an attribute error. This change puts in a
dummy value (1 January 1970) for the built field.

```
$ gbp diff babette 104 1130
diff -r babette/104 babette/1130
Traceback (most recent call last):
  File "/home/marduk/.local/bin/gbp", line 8, in <module>
    sys.exit(main())
             ^^^^^^
  File "/home/marduk/.local/pipx/venvs/gbpcli/lib/python3.12/site-packages/gbpcli/__init__.py", line 406, in main
    return cast(int, args.func(args, GBP(args.url), console))
                     ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/home/marduk/.local/pipx/venvs/gbpcli/lib/python3.12/site-packages/gbpcli/subcommands/diff.py", line 41, in handler
    header(f"--- a/{args.machine}/{left} {render.timestr(left_build.info.built)}")
                                          ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/home/marduk/.local/pipx/venvs/gbpcli/lib/python3.12/site-packages/gbpcli/render.py", line 27, in timestr
    dt_obj = timestamp.astimezone(timezone)
             ^^^^^^^^^^^^^^^^^^^^
AttributeError: 'NoneType' object has no attribute 'astimezone'
```