allow using config file from a custom location
I Perhaps should have forseen this. By default a config file is checked
in ~/.config/gbpcli.toml (or whatever config directory for the
respective platform). However if the environment variable
`GBPCLI_CONFIG` file is defined, the referenced file is used instead. In
this case if the file does not exist an error is raised.