packages: fix case when build doesn't exist
When retrieving packages from a build that does not exist, the GraphQL
response has, e.g.

    {"data": {"build": None}}

However the .packages() method is assuming that "build" is populated
(with an empty list of packages), resulting in a "NoneType object has no
attribute .get" error.

This change fixes .packages() to handle the actual GraphQL result.