add BaseSettings from gentoo-build-publisher
There are plugins that (want to) use BaseSettings, however it is defined
in gentoo-build-publisher, so they must pull in all of gbp. Since a lot
of these have cli components we don't want to have to pull all of gbp in
just to import Settings (which actually might not even be used by the
cli).  Therefore must move it into gbpcli and have Gentoo Build
Publisher use that (gentoo-build-publisher already depends on gbpcli).