settings: allow custom validator methods
This change allows it so that settings fields can have their own
validator methods, similar to the way Django forms, DRF, and many other
libraries use. A typical example would be method to convert a string to
a datetime object.

This was always intended to be a part of `Settings` but somehow it fell
through the cracks. Now it has been implemented.