utils.resolve_build_id: xlate "@@" to latest build
For a long time I have wanted something like this. "@foo" means a build
with the tag "foo". "@" means the published tag. But there was no way to
specify the **latest** build for a machine other than, e.g.:

    $ gbp notes babette $(gbp latest babette)

With this change, utils.resolve_build_id() recognizes "@@" to mean
"latest build" and calls latest for you. So the above command line can
be replaced with:

    $ gbp notes babette @@

Note that "@@" gets translated client-side. The server does not
recognize "@@" as a special tag meaning latest. In fact currently on the
server "@" is a valid tag name, which needs to be fixed.