diff: don't compile regex at import time
This is a minor perf tweak so that the CPV regex isn't compiled at
import time. Rather we call re.match() in the function that uses the
regex which Python silently compiles and caches.

This is so that, e.g. `gbp --help` doesn't waste time compiling a regex
that it isn't going to use.